Prepare for variable location of node_modules
To prepare to run the development environment from a set of services in
docker-compose, we need to perform the "node_modules volume trick" which
will allow having a `node_modules` directory on the host and another one
inside the docker image, without interference. The docker image will
have its `node_modules` directory in a parent directory, a sibling to
the project directory.

This means the app needs to ready to have `node_modules` in a location
that is not necessarily within the project directory.

This changeset introduces the following adjustments to achieve that:

1. `gulp/config.js` exposes the path to the `node_modules` where
`fontawesome-free` is installed. The rest of *gulp* tasks use that
instead of hard-coded paths.

2. A new module, `node-sass-tilde-importer`, is introduced. It allows
importing in SASS with a tilde prefix (`~`) which will resolve to the
currently used `node_modules` directory.