Dev docker compose (#317)
* Prepare for variable location of node_modules

To prepare to run the development environment from a set of services in
docker-compose, we need to perform the "node_modules volume trick" which
will allow having a `node_modules` directory on the host and another one
inside the docker image, without interference. The docker image will
have its `node_modules` directory in a parent directory, a sibling to
the project directory.

This means the app needs to ready to have `node_modules` in a location
that is not necessarily within the project directory.

This changeset introduces the following adjustments to achieve that:

1. `gulp/config.js` exposes the path to the `node_modules` where
`fontawesome-free` is installed. The rest of *gulp* tasks use that
instead of hard-coded paths.

2. A new module, `node-sass-tilde-importer`, is introduced. It allows
importing in SASS with a tilde prefix (`~`) which will resolve to the
currently used `node_modules` directory.

* Add Dockerfile to run gulp watch

* Add Dockerfile to run the main app

* Add docker-compose services

This change introduces a docker-compose.yml file defining three
services, as extracted from the README instructions:

1. A postgres instance with a named volume so that data can be persisted
across runs

2. A gulp service that watches changes in frontend files and regenerates
assets

3. A pycan_web service with the main Django app running on the
development server.

Note that prior to execution the migrations need to be run:

    docker-compose run pycan_web ./manage.py migrate

And after that all services can be started with:

    docker-compose up

* Fix Environment variable in Docker file

* Expose postgres ports to the outside

So one is able to connect and debug the DB.

* Make the media directory a data volume

So it persists across restarts.

* Add dev management command to add initial data for development

Run manage.py dbload to load a couple of things that will make the web
usable by a developer first starting.

Photo by https://unsplash.com/@srosinger3997

* Describe docker-compose behaviour in README

* Fix anchor in README's TOC

* Ensure directories are treated as such in docker copy commands

* Reorder docker-compose instructions for clarity

Leave the command that blocks the terminal for last.

* Disable pip version warnings in docker build

Pip comes installed with the base docker image, we are not planning to
upgrade here, the warning just adds noise in the docker build step.

* Allow changing the database and app host port binds

Through environment variables.

Co-authored-by: Jacobo de Vera <jacobo.devera@hotjar.com>